requirements.txt
python-rtmidi==1.2.1
midiutil==1.2.1
requests==2.21.0
slackclient==1.3.0
midi2audio==0.1.1
mido==1.2.9
oauthlib==3.0.0
google-auth-oauthlib==0.2.0
requests-oauthlib==1.2.0